test: fixing and adding tests and minor fixes
