+++ oracle-api/src/main/java/ca/bc/gov/oracleapi/endpoint/consep/MoistureContentConesEndpoint.java
@@ -216,7 +216,7 @@ public class MoistureContentConesEndpoint {
       activityData.setActualBeginDateTime(moistureContent.get().actualBeginDateTime());
       activityData.setActualEndDateTime(moistureContent.get().actualEndDateTime());
       activityData.setTestCategoryCode(moistureContent.get().testCategoryCode());
-      activityService.validateMoistureContentActivityData(activityData);
+      activityService.validateActivityData(activityData);
       testResultService.updateTestResultStatusToCompleted(riaKey);
 
     } catch (Exception e) {
